Update user info enum
