+++ src/main/java/oneus/GSMATCH/domain/user/controller/UserController.java
@@ -6,6 +6,8 @@ import lombok.RequiredArgsConstructor;
 import oneus.GSMATCH.domain.user.dto.request.LoginRequest;
 import oneus.GSMATCH.domain.user.dto.request.SignOutRequest;
 import oneus.GSMATCH.domain.user.dto.request.SignupRequest;
+import oneus.GSMATCH.domain.user.entity.UserEntity;
+import oneus.GSMATCH.domain.user.repository.UserRepository;
 import oneus.GSMATCH.domain.user.service.UserService;
 import oneus.GSMATCH.global.security.UserDetailsImpl;
 import oneus.GSMATCH.global.util.MsgResponseDto;
@@ -14,11 +16,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/auth")
 public class UserController {
     private final UserService userService;
+    private final UserRepository userRepository;
 
     // 회원 가입
     @PostMapping("/signup")
@@ -39,5 +44,10 @@ public class UserController {
     private ResponseEntity<MsgResponseDto> signOut(@RequestBody SignOutRequest signOutRequestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
         userService.signOut(signOutRequestDto, userDetails.getUser());
         return ResponseEntity.ok(new MsgResponseDto("회원탈퇴 완료", HttpStatus.OK.value()));
+    }
+
+    @GetMapping("/a")
+    public ResponseEntity<List<UserEntity>> a() {
+        return ResponseEntity.ok(userRepository.findAll());
     }
 }
