+++ src/main/java/oneus/GSMATCH/domain/user/service/UserService.java
@@ -59,8 +59,8 @@ public class UserService {
                 .gender(signupRequestDto.getGender())
                 .major(signupRequestDto.getMajor())
                 .type(signupRequestDto.getType())
-                .level(0)
-                .point(0)
+                .level(1) // default 레벨 1
+                .point(0) // default 포인트 0
                 .build();
 
         userRepository.save(user);
