+++ src/main/java/oneus/GSMATCH/global/util/UserStateEnum.java
@@ -1,0 +1,33 @@
+package oneus.GSMATCH.global.util;
+
+public class UserStateEnum {
+    public enum Grade {
+        ONE,
+        TWO,
+        THREE;
+    }
+
+    public enum Gender {
+        MALE,
+        FEMALE;
+    }
+
+    public enum Major {
+        FRONT,
+        BACK,
+        DESIGN,
+        IOS,
+        AOS,
+        DEVOPS,
+        WORLD_SKILL,
+        IOT,
+        AI;
+    }
+
+    public enum Type {
+        PORORO,
+        LUPI,
+        POBI,
+        EDI;
+    }
+}
